+++ src/cli/mod.rs
@@ -1,5 +1,6 @@
 mod action;
 mod args;
+mod config;
 mod help;
 mod instance;
 mod port;
@@ -10,4 +11,4 @@ mod run;
 //                                   Public                                  //
 //===========================================================================//
 
-pub use self::{action::*, args::*, instance::*, run::*};
+pub use self::{action::*, args::*, config::*, instance::*, run::*};